refactor(rust): extract git.rs with 14 git operation commands
